+++ hydra-node/src/Hydra/HeadLogic.hs
@@ -101,12 +101,13 @@ import Hydra.Tx.Snapshot (ConfirmedSnapshot (..), Snapshot (..), SnapshotNumber,
 defaultTTL :: TTL
 defaultTTL = 5
 
+-- TODO: refine semantics with overall 'Connectivity'
 onConnectionEvent :: Connectivity -> Outcome tx
 onConnectionEvent = \case
-  Connected{nodeId} ->
-    causes [ClientEffect (ServerOutput.PeerConnected nodeId)]
-  Disconnected{nodeId} ->
-    causes [ClientEffect (ServerOutput.PeerDisconnected nodeId)]
+  Connected{} ->
+    causes [ClientEffect ServerOutput.NetworkConnected]
+  Disconnected{} ->
+    causes [ClientEffect ServerOutput.NetworkDisconnected]
   HandshakeFailure{remoteHost, ourVersion, theirVersions} ->
     causes
       [ ClientEffect